Added FREEZEOUT_VH macro to convert vah freezeout surface to standard vh format

+++ rhic/include/Memory.h
@@ -4,11 +4,11 @@
 
 double *** calloc_3d_array(double ***array, int dim1, int dim2, int dim3);
 double **** calloc_4d_array(double ****array, int dim1, int dim2, int dim3, int dim4);
-double ***** calloc_5d_array(double *****array, int dim1, int dim2, int dim3, int dim4, int dim5);
+float ***** calloc_5d_array(float *****array, int dim1, int dim2, int dim3, int dim4, int dim5);
 
 void free_3d_array(double ***array, int dim1, int dim2);
 void free_4d_array(double ****array, int dim1, int dim2, int dim3);
-void free_5d_array(double *****array, int dim1, int dim2, int dim3, int dim4);
+void free_5d_array(float *****array, int dim1, int dim2, int dim3, int dim4);
 
 
 #endif
